+++ src/main/java/com/comanda/model/form/EstoqueMovimentoFormR.java
@@ -1,12 +1,13 @@
 package com.comanda.model.form;
 
+import java.math.BigDecimal;
+
 import com.comanda.domain.entity.EstoqueMovimento;
-import com.comanda.domain.entity.Produto;
 import com.comanda.domain.enumerado.TipoMovimentacao;
 
 public record EstoqueMovimentoFormR( 
 		TipoMovimentacao tipo, ProdutoFormR produto,
-				 Integer qtde, Integer saldoanterior) {
+				 BigDecimal qtde, BigDecimal saldoanterior) {
 	
 	public EstoqueMovimentoFormR(EstoqueMovimento estoqueMovimento) {
 		this(estoqueMovimento.getTipoMovimentacao(),new ProdutoFormR(estoqueMovimento.getProduto().getId()),
