+++ src/main/java/com/comanda/model/recorddto/EstoqueMovimentoDToR.java
@@ -1,10 +1,12 @@
 package com.comanda.model.recorddto;
 
+import java.math.BigDecimal;
+
 import com.comanda.domain.entity.EstoqueMovimento;
 import com.comanda.domain.enumerado.TipoMovimentacao;
 
 public record EstoqueMovimentoDToR(TipoMovimentacao tipoMovimentacao, 
-		 Integer qtde, Integer saldoanterior) {
+		 BigDecimal qtde, BigDecimal saldoanterior) {
 	
 	public EstoqueMovimentoDToR(EstoqueMovimento estoqueMovimento) {
 		this(estoqueMovimento.getTipoMovimentacao(), estoqueMovimento.getQtde(), estoqueMovimento.getSaldoanterior());
